dsa\lab4: make time more precise

+++ dsa/lab4/common/time.c
@@ -1,10 +1,10 @@
 #include <time.h>
 
-double current_time() {
+long double current_time() {
 #if __STDC_VERSION__ >= 201112L  // C11
     struct timespec ts;
     timespec_get(&ts, TIME_UTC);
-    return (double)ts.tv_sec + (double)ts.tv_nsec / 1e+9;
+    return (long double)ts.tv_sec + (long double)ts.tv_nsec / 1e+9;
 #else
     return clock() / CLOCKS_PER_SEC;
 #endif
