+++ dsa/lab4/common/time.h
@@ -1,8 +1,8 @@
-double current_time();
+long double current_time();
 
 #define TIMEIT(timing, code)              \
     {                                     \
-        double begin = current_time();    \
+        long double begin = current_time();    \
         code;                             \
         timing += current_time() - begin; \
     }
